backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.22
django-braces==1.17.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.15.4
httpie==3.2.4
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.4.3
pillow==11.1.0
pipenv==2024.0.1
platformdirs==4.2.2
Pmw==2.1.1
psycopg2==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.17.0
PySocks==1.7.1
python-dotenv==1.1.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
Werkzeug==3.0.4
zipp==3.20.1
